feat: system-consolidation-validation-test-file-exemption-burn-down-001 - Preserve canonical finding and stable identity validation evidence
